# services/tps/api/generated/integrations_models.py
# generated by datamodel-codegen:
#   filename:  integrations.yaml
#   timestamp: 2026-04-11T06:17:17+00:00

from __future__ import annotations

from typing import Annotated, Any

from pydantic import BaseModel, Field


class IntegrationDTO(BaseModel):
    id: str
    app_name: str
    identifier: str | None = None
    status: str
    created_at: str


class InstallRequest(BaseModel):
    form_data: Annotated[
        dict[str, Any] | None,
        Field(
            description='For form-based OAuth2 — user-supplied fields (e.g., tenant URL)'
        ),
    ] = None


class InstallResponse(BaseModel):
    authorize_url: str
    state: str


class ExchangeRequest(BaseModel):
    code: str
    state: str


class ConnectRequest(BaseModel):
    credentials: Annotated[
        dict[str, Any], Field(description='Fields matching app.meta.form_fields')
    ]


class ConnectResponse(BaseModel):
    integration_id: str
    app_name: str
    identifier: str | None = None
    status: str


class ResolveStateResponse(BaseModel):
    app_name: str


class CloneRequest(BaseModel):
    repo_url: str


class CloneResponse(BaseModel):
    clone_dir: str
    owner: str
    repo: str


class OkResponse(BaseModel):
    ok: bool